+++ R/auto-clean.R
@@ -3,12 +3,13 @@
 #' @description
 #' Perform automatic data preprocessing on glycoproteomics or glycomics data.
 #' This function applies an intelligent preprocessing pipeline that includes
-#' normalization, missing value handling, imputation, aggregation (for
-#' glycoproteomics data), and batch effect correction.
+#' normalization, missing value handling, imputation, aggregation, and batch
+#' effect correction.
 #'
 #' For glycomics data, this function calls these functions in sequence:
 #' - [auto_remove()]
 #' - [auto_impute()]
+#' - [aggregate()]
 #' - [auto_normalize()]
 #' - [auto_correct_batch_effect()]
 #'
@@ -16,7 +17,7 @@
 #' - [auto_remove()]
 #' - [auto_normalize()]
 #' - [auto_impute()]
-#' - [auto_aggregate()]
+#' - [aggregate()]
 #' - [auto_normalize()]
 #' - [auto_correct_batch_effect()]
 #'
@@ -49,7 +50,8 @@
 #' exp <- real_experiment
 #' auto_clean(exp)
 #'
-#' @seealso [auto_normalize()], [auto_remove()], [auto_impute()], [auto_aggregate()], [auto_correct_batch_effect()]
+#' @seealso [auto_normalize()], [auto_remove()], [auto_impute()], [aggregate()],
+#'   [auto_correct_batch_effect()]
 #' @export
 auto_clean <- function(
   exp,
@@ -116,7 +118,7 @@ auto_clean <- function(
   cli::cli_alert_success("Imputation completed.")
 
   cli::cli_h2("Aggregating data")
-  exp <- auto_aggregate(exp, standardize_variable = params$standardize_variable)
+  exp <- aggregate(exp, standardize_variable = params$standardize_variable)
   cli::cli_alert_success("Aggregation completed.")
 
   cli::cli_h2("Normalizing data again")
@@ -154,6 +156,10 @@ auto_clean <- function(
     params$group_col
   )
   cli::cli_alert_success("Imputation completed.")
+
+  cli::cli_h2("Aggregating data")
+  exp <- aggregate(exp, standardize_variable = params$standardize_variable)
+  cli::cli_alert_success("Aggregation completed.")
 
   cli::cli_h2("Normalizing data")
   exp <- auto_normalize(